Refactor profile and prompt fetching to REST
