+++ App/services/challengeStreakService.ts
@@ -1,5 +1,5 @@
-import { getDocument, setDocument } from '@/services/firestoreService';
-import { updateUserProfile } from '../../utils/firestoreHelpers';
+
+import { updateUserProfile, loadUserProfile } from '../../utils/userProfile';
 import { getCurrentUserId } from '@/utils/TokenManager';
 import { ensureAuth } from '@/utils/authGuard';
 
@@ -7,7 +7,7 @@ export async function completeChallengeWithStreakCheck(): Promise<number | null>
   const userId = await ensureAuth(await getCurrentUserId());
   if (!userId) return null;
 
-  const userData = await getDocument(`users/${userId}`);
+  const userData = await loadUserProfile(userId);
   const streakData = userData?.challengeStreak || {};
   const currentCount = streakData.count || 0;
   const lastCompletedDate = streakData.lastCompletedDate
