+++ App/services/onboardingService.ts
@@ -1,6 +1,5 @@
 import { navigationRef } from '@/navigation/navigationRef';
-import { setDocument, getDocument } from '@/services/firestoreService';
-import { updateUserProfile } from '../../utils/firestoreHelpers';
+import { loadUserProfile, updateUserProfile } from '../../utils/userProfile';
 import { ensureAuth } from '@/utils/authGuard';
 
 export async function saveUsernameAndProceed(username: string): Promise<void> {
@@ -13,7 +12,7 @@ export async function saveUsernameAndProceed(username: string): Promise<void> {
 
 export async function checkIfUserIsNewAndRoute(): Promise<void> {
   const uid = await ensureAuth();
-  const profile = await getDocument(`users/${uid}`);
+  const profile = await loadUserProfile(uid);
   const completed = !!profile?.onboardingComplete;
   if (navigationRef.isReady()) {
     const initialRoute = completed ? 'Home' : 'Onboarding';
