+++ App/utils/TokenManager.ts
@@ -1,5 +1,5 @@
 import { getDocument, setDocument } from '@/services/firestoreService';
-import { updateUserProfile } from '../../utils/firestoreHelpers';
+import { updateUserProfile } from '../../utils/userProfile';
 import { ensureAuth } from '@/utils/authGuard';
 
 export const getTokenCount = async () => {
